+++ eslint.preset.mjs
@@ -50,6 +50,7 @@ export default [
   {
     ignores: [
       '**/dist/**',
+      '**/lib/**',
       '**/node_modules/**',
       '**/.angular/**',
       '.wireit/**',